[MOVE/#8] 파일 구조화
