+++ app/src/main/java/umc/everyones/lck/presentation/lck/AboutLckTeamPlayerFragment.kt
@@ -9,6 +9,8 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import umc.everyones.lck.R
+import umc.everyones.lck.presentation.lck.adapter.PlayerCareerAdapter
+import umc.everyones.lck.presentation.lck.data.PlayerCareerData
 
 class AboutLckTeamPlayerFragment : Fragment() {
