--- app/src/main/java/umc/everyones/lck/presentation/lck/OnPlayerItemClickListener.kt
@@ -1,5 +1,0 @@
-package umc.everyones.lck.presentation.lck
-
-interface OnPlayerItemClickListener {
-    fun onPlayerItemClick(player: PlayerData)
-}
