--- app/src/main/java/umc/everyones/lck/presentation/lck/OnTeamClickListener.kt
@@ -1,5 +1,0 @@
-package umc.everyones.lck.presentation.lck
-
-interface OnTeamClickListener {
-    fun onTeamClick(team: RankingData)
-}
