+++ app/src/main/java/umc/everyones/lck/presentation/lck/adapter/MatchDetailsAdapter.kt
@@ -1,4 +1,4 @@
-package umc.everyones.lck.presentation.lck
+package umc.everyones.lck.presentation.lck.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -7,6 +7,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import umc.everyones.lck.R
+import umc.everyones.lck.presentation.lck.data.MatchData
 
 class MatchDetailsAdapter(private val matchDetails: List<MatchData>) :
     RecyclerView.Adapter<MatchDetailsAdapter.MatchDetailViewHolder>() {
