+++ app/src/main/java/umc/everyones/lck/presentation/lck/adapter/MatchVPAdapter.kt
@@ -1,4 +1,4 @@
-package umc.everyones.lck.presentation.lck
+package umc.everyones.lck.presentation.lck.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import umc.everyones.lck.R
+import umc.everyones.lck.presentation.lck.data.MatchData
 
 
 class MatchVPAdapter : RecyclerView.Adapter<MatchVPAdapter.MatchViewHolder>() {
