+++ app/src/main/java/umc/everyones/lck/presentation/lck/data/HistoryData.kt
@@ -1,4 +1,4 @@
-package umc.everyones.lck.presentation.lck
+package umc.everyones.lck.presentation.lck.data
 
 data class HistoryData(
     val title: String,
