+++ app/src/main/java/umc/everyones/lck/presentation/lck/data/PlayerData.kt
@@ -1,4 +1,4 @@
-package umc.everyones.lck.presentation.lck
+package umc.everyones.lck.presentation.lck.data
 
 data class PlayerData(
     val playerImg: Int,
