+++ app/src/main/java/umc/everyones/lck/presentation/lck/util/CustomRecyclerView.kt
@@ -1,5 +1,5 @@
 
-package umc.everyones.lck.presentation.lck
+package umc.everyones.lck.presentation.lck.util
 
 import android.content.Context
 import android.util.AttributeSet
