+++ app/src/main/java/umc/everyones/lck/presentation/lck/util/OnPlayerItemClickListener.kt
@@ -1,0 +1,7 @@
+package umc.everyones.lck.presentation.lck.util
+
+import umc.everyones.lck.presentation.lck.data.PlayerData
+
+interface OnPlayerItemClickListener {
+    fun onPlayerItemClick(player: PlayerData)
+}
