+++ app/src/main/java/umc/everyones/lck/presentation/lck/util/OnTeamClickListener.kt
@@ -1,0 +1,7 @@
+package umc.everyones.lck.presentation.lck.util
+
+import umc.everyones.lck.presentation.lck.data.RankingData
+
+interface OnTeamClickListener {
+    fun onTeamClick(team: RankingData)
+}
