+++ app/src/main/res/layout/item_about_lck_player.xml
@@ -43,17 +43,20 @@
             app:layout_constraintBottom_toBottomOf="@id/tv_about_lck_winning_career"
             app:layout_constraintEnd_toEndOf="parent"/>
 
-        <androidx.recyclerview.widget.RecyclerView
+        <umc.everyones.lck.presentation.lck.util.CustomRecyclerView
             android:id="@+id/rv_about_lck_player_detail"
-            android:layout_width="wrap_content"
-            android:layout_height="0dp"
-            android:maxHeight="10dp"
+            android:layout_width="match_parent"
+            android:layout_height="110dp"
             android:nestedScrollingEnabled="true"
             android:scrollbars="vertical"
+            android:scrollbarThumbVertical="@drawable/custom_scrollbar_thumb"
+            android:scrollbarTrackVertical="@drawable/custom_scrollbar_track"
             android:layout_marginTop="20dp"
-            android:layout_marginStart="8dp"
+            android:layout_marginStart="37dp"
+            android:layout_marginEnd="40dp"
             app:layout_constraintTop_toBottomOf="@id/tv_about_lck_winning_career"
-            app:layout_constraintStart_toStartOf="@id/view_about_lck_team_player_line1"/>
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"/>
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
